song with cover and rating work

+++ backend/src/main/java/ro/iss/backend/domain/Song.java
@@ -12,7 +12,7 @@ public class Song {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "song_id")
-    private Long songId;
+    private int songId;
 
     @Column(name = "song_name")
     private String songName;
